move implementation of Schema__*() in schema.cpp
